requirements.txt
streamlit>=1.30
pytest>=7.0
google-generativeai>=0.7.2
python-dotenv>=1.0.1
